// frontend/src/interfaces/IHeadphoneAudioExposure.ts
export interface IHeadphoneAudioExposure {
  _id: string;
  userID: string;
  startDate: Date;
  duration: number;
  value: number;
  hkID: string;
}

export interface IHeadphoneAudioExposureDTO {
  userID: string;
  startDate: Date;
  duration: number;
  value: number;
  hkID: string;
}

export interface IHeadphoneAudioExposureLocal{
  userID: string;
  startDate: Date;
  value: number;

}